requirements.txt
numpy==1.17.2
Pillow==9.3.0
scipy==1.1.0
torch==1.10.0
torchvision==0.4.1
